Выполнил задание 2
