+++ reserve/.stylelintrc
@@ -1,0 +1,3 @@
+{
+  "extends": "stylelint-config-htmlacademy"
+}
