+++ src/components/films-list.js
@@ -1,0 +1,11 @@
+const createFilmsListTemplate = () => {
+  return (
+    `<section class="films">
+      <section class="films-list">
+        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+      </section>
+    </section>`
+  );
+};
+
+export {createFilmsListTemplate};
