+++ src/components/list-container.js
@@ -1,0 +1,7 @@
+const createListContainerTemplate = () => {
+  return (
+    `<div class="films-list__container"></div>`
+  );
+};
+
+export {createListContainerTemplate};
