+++ src/components/list-extra.js
@@ -1,0 +1,9 @@
+const createListExtraTemplate = () => {
+  return (
+    `<section class="films-list--extra">
+      <h2 class="films-list__title"></h2>
+    </section>`
+  );
+};
+
+export {createListExtraTemplate};
