+++ src/components/statistic.js
@@ -1,0 +1,7 @@
+const createStatisticTemplate = () => {
+  return (
+    `<p>130 291 movies inside</p>`
+  );
+};
+
+export {createStatisticTemplate};
